feat(membrane): update runtime-math kernels and membrane validation for expanded ABI
Update the Transparent Safety Membrane layer to support expanded native
implementations:

- arena/ptr_validator: validation adjustments for new ABI entrypoints
- runtime_math kernels (derived_tstructure, info_geometry, kernel_mmd,
  operator_norm, spectral_gap, stein_discrepancy, transfer_entropy,
  wasserstein_drift): improved numerical stability and cross-family
  consistency checks across all runtime-math control kernels

These changes ensure the membrane's online decision-making and
deterministic admission gates remain coherent with the expanded set of
natively-implemented libc symbols.

+++ crates/frankenlibc-membrane/src/runtime_math/derived_tstructure.rs
@@ -183,7 +183,7 @@ impl TStructureController {
                 let mut worst: Option<u8> = None;
                 for d in (0..degree).rev() {
                     if self.completed_mask & (1u8 << d) == 0 {
-                        worst = Some(d as u8);
+                        worst = Some(d);
                         break;
                     }
                 }
